requirements.txt
discord.py
xlsxwriter
os
time
